Code Correction

+++ src/main/resources/db/migration/V01__create_ocurrences_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE ocurrences (
 	codigo BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
 	datatime DATETIME,
-	number int(100) NOT NULL
+	number BIGINT(100) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
